requirements.txt
apiview==1.3.25
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-import-export==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
Faker==4.1.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pycountry==19.8.18
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
xlrd==1.2.0
xlwt==1.3.0
